backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
# psycopg2-binary - NOT NEEDED (using Supabase client instead)
alembic==1.12.1

# Supabase
supabase==2.3.0

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
joblib==1.3.2

# Google Gemini AI
google-generativeai==0.3.1

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
faker==20.1.0

# HTTP & File Handling
httpx==0.25.2
pillow==10.1.0
aiofiles==23.2.1

# Logging & Monitoring
loguru==0.7.2

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
